+++ lib/macos_ui.dart
@@ -20,3 +20,5 @@ export 'src/util.dart';
 export 'src/styles/theme.dart';
 export 'src/styles/typography.dart';
 export 'src/layout/scaffold.dart';
+export 'src/buttons/switch.dart';
+export 'src/indicators/progress_indicators.dart';